+++ src/main/java/config/Setting.java
@@ -1,0 +1,6 @@
+package config;
+
+public class Setting {
+    public static final String FILE_DIRECTORY = "./data";
+    public static final String FILE_NAME = "tasks.txt";
+}
