+++ src/main/java/models/Event.java
@@ -1,4 +1,6 @@
-package processes;
+package models;
+
+import models.Task;
 
 public class Event extends Task {
