+++ src/main/java/processor/IProcessor.java
@@ -1,4 +1,4 @@
-package processes;
+package processor;
 
 import java.util.List;
